+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 2.17.0 (2023-10-14)
 
 * (theimo1221) Add Tuya device handling
 * (theimo1221) Add first Tuya device (Tuya Garage Door Opener)